delete function bug fix

+++ backend/controller/categoryController.js
@@ -4,6 +4,7 @@ const path = require('path');
 
 const addCategory = async (req, res) => {
     try {
+        if (!(req.uRole === "admin")) return res.status(500).send({ err: "Server is down!" });
         let productData = req.body;
 
         if (req.files && req.files.length > 0) {
@@ -28,12 +29,12 @@ const getAllCategory = async (req, res) => {
 
     try {
 
-        const products = await Category.find();
+        const category = await Category.find();
 
-        if (!products) {
+        if (!category) {
             return res.status(404).json({ err: "False Attempted!" });
         }
-        res.status(200).json({ products: products });
+        res.status(200).json({ category: category });
 
     } catch (error) {
 
@@ -57,6 +58,7 @@ const getCategory = async (req, res) => {
 }
 const updateCategory = async (req, res) => {
     try {
+        if (!(req.uRole === "admin")) return res.status(500).send({ err: "Server is down!" });
         const { id, ...bodyData } = { ...req.body };
 
         const category = await Category.findByIdAndUpdate(id, bodyData);
@@ -65,7 +67,7 @@ const updateCategory = async (req, res) => {
                 err: "Server is down!"
             });
         }
-        res.status(200).json({ mess: "You got a update!" });
+        res.status(200).json({ message: "You got an update!" });
     } catch (error) {
         res.status(500).send({
             err: "Bad request!"
@@ -74,16 +76,17 @@ const updateCategory = async (req, res) => {
 }
 const removeCategory = async (req, res) => {
     try {
+        if (!(req.uRole === "admin")) return res.status(500).send({ err: "Server is down!" });
         const id = req.body.id;
         const category = await Category.findByIdAndDelete(id).select('image -_id');
 
         if (!category) {
             return res.status(404).send({
-                err: "Ser ver is down!"
+                err: "Server is down!"
             });
         }
         const fileName = category.image;
-        if (!(fileName === "default-avatar.png")) {
+        if (!(fileName === "default-product.png")) {
             const fileDest = '../public/uploads/categories/';
 
             fs.unlink(path.join(__dirname, fileDest + fileName), (err) => {
@@ -92,7 +95,7 @@ const removeCategory = async (req, res) => {
                 }
             });
         }
-        res.status(200).json({ mess: "Deleted Successfully!" });
+        res.status(200).json({ message: "Deleted Successfully!" });
     } catch (error) {
         res.status(500).send({
             err: "Bad Request!"
